Fixes to keyboard code, making PC-compatible initialization working. (#1)
* Update esp32-ps2dev.cpp

The interface was missing some commands in one of my sytems because of waiting to send the 0xAA "BAT test success". BAT success should always be sent a couple of houndred of milliseconds after power up, regarding state.

* Update README.md

* Update esp32-ps2dev.cpp

Testing in my Toshiba 205CDS with a combined port, DEVICE ID command was failing to write the ID properly. Added some loops to shield time to the controller and get them properly.

* Update esp32-ps2dev.cpp

On my 486 PC, ACKs were being missed by the host controller during LED set command handling. This made the process to end and the controller to ignore the keyboard after that command. This makes it stable.

* Added delay on reset command ack